requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
PyYAML>=6.0
Flask>=3.0.0
qrcode>=7.4
jieba>=0.42.1
snownlp>=0.12.3
